requirements_full.txt
# AIBET Analytics Platform - Complete Dependencies
# AI/ML Stack
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.25.2
joblib>=1.3.2

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Telegram Bot
aiogram>=3.4.1
aiohttp>=3.9.0
aiohttp-cors>=0.7.0

# Database
aiosqlite>=0.19.0
sqlalchemy>=2.0.23

# Data Collection/Parsing
beautifulsoup4>=4.12.2
lxml>=4.9.3
requests>=2.31.0
httpx>=0.25.2
selenium>=4.15.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
pytz>=2023.3
tenacity>=8.2.3
cachetools>=5.3.2
async-limiter>=0.1.1

# Security
cryptography>=41.0.8

# Performance
psutil>=5.9.6

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Visualization (for reports)
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.17.0

# Additional HTTP clients
httpcore>=1.0.0
pydantic-core>=2.14.5
anyio>=4.0.0
